Asset : Controller + Model + Service Implemented

+++ src/main/java/com/eams/user/repository/UserRepository.java
@@ -2,9 +2,10 @@ package com.eams.user.repository;
 
 import com.eams.user.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
     boolean existsByEmail(String email);
 }
